requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.3
html-json-forms==1.0.0
idna==2.6
PyJWT==1.6.1
pytz==2018.4
requests==2.18.4
urllib3==1.22
